+++ tests/test_needs_builder.py
@@ -37,3 +37,35 @@ def test_doc_needs_build_without_needs_file(app, status, warning):
 @with_app(buildername="needs", srcdir="../docs")
 def test_needs_official_doc(app, status, warning):
     app.build()
+
+
+@with_app(buildername="html", srcdir="doc_test/doc_needs_builder_parallel")
+def test_needs_html_and_json(app, status, warning):
+    """
+    Build html output and needs.json in one sphinx-build
+    """
+    import json
+    import os
+    import subprocess
+
+    app.build()
+
+    needs_json_path = os.path.join(app.outdir, "needs.json")
+    assert os.path.exists(needs_json_path)
+
+    srcdir = "doc_test/doc_needs_builder_parallel"
+    build_dir = os.path.join(app.outdir, "../needs")
+    subprocess.run(["sphinx-build", "-b", "needs", srcdir, build_dir], capture_output=True)
+    needs_json_path_2 = os.path.join(build_dir, "needs.json")
+    assert os.path.exists(needs_json_path_2)
+
+    # Check if the needs.json files from html/parallel build and builder are the same
+    with open(needs_json_path) as f1:
+        needs_1 = json.load(f1)
+        with open(needs_json_path_2) as f2:
+            needs_2 = json.load(f2)
+
+            # Just check need-data, as the rest contains not matching timestamps
+            need_data_1 = needs_1["versions"]["1.0"]["needs"]
+            need_data_2 = needs_2["versions"]["1.0"]["needs"]
+            assert need_data_1 == need_data_2